+++ server/app/routes.py
@@ -13,9 +13,10 @@ from nmtwizard import common, task
 from nmtwizard.helper import build_task_id, shallow_command_analysis, change_parent_task, remove_config_option
 
 logger = logging.getLogger(__name__)
-logger.addHandler(ch)
+logger.addHandler(app.logger)
 
 def get_service(service):
+    app.logger.info("Hello world!")
     """Wrapper to fail on invalid service."""
     def_string = redis.hget("admin:service:"+service, "def")
     if def_string is None: